refactor(civitai): move the request pacer into the Civitai project
The gateway that will own pacing lives in DiffusionNexus.Civitai, which
cannot reference Service. WaitAsync gains a per-call interval so one pacer
serves both lanes off a single shared timestamp.
